Add fractional retrieval counts for passive surfacing; bump to v1.21.1
Per-prompt search (semantic_context) now weakly increments retrieval_count
by a configurable fraction (default 0.01) for passively surfaced Tier 2
documents. Explicit /recall still increments by 1.0. This creates a slow
accumulation path so frequently-surfaced observations can eventually reach
the promotion threshold (>3) without manual recall.

Changes:
- int→float migration for retrieval_count across query, tier2, promotion, scoring
- _increment_retrieval_counts() accepts increment param with round(,2) precision
- semantic_context() calls fractional increment after dedup/threshold filtering
- New config key: memory.per_prompt_search.passive_retrieval_increment (0.01)
- Promotion display format shows 2 decimal places for retrieval counts
- 11 new tests (972 total), all passing

Version bump: 1.21.0 → 1.21.1 (patch)

+++ plugins/jarvis/mcp-server/tools/scoring.py
@@ -42,7 +42,7 @@ def compute_importance(
     vault_type: str = "unknown",
     frontmatter_importance: Optional[str] = None,
     created_at: Optional[str] = None,
-    retrieval_count: int = 0,
+    retrieval_count: float = 0,
     config: Optional[dict] = None,
 ) -> float:
     """Compute importance score for a document.
@@ -113,7 +113,7 @@ def _compute_recency_bonus(created_at: Optional[str], half_life_days: float = 7.
         return 0.0
 
 
-def _compute_retrieval_bonus(retrieval_count: int) -> float:
+def _compute_retrieval_bonus(retrieval_count: float) -> float:
     """Log-scaled bonus from retrieval count.
 
     log2(count + 1) / 10, capped at _RETRIEVAL_BONUS_MAX.
